# OpenAPI/coffeeshops.yaml
openapi: 3.1.1
info:
  title: CoffeeShops
  description: REST API для системы управления кофейнями, напитками и пользователями
  version: 1.0.0
  contact: {}
servers:
  - url: /api/v1
    description: Основной сервер API
    
#-----------------------------------------SCHEMAS-----------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        user_id:
          type: string
          example: "a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6"
    CreateResponse:
      type: object
      properties:
        id:
          type: string
          example: "9b165097-1c9f-4ea3-bef0-e505baa4ff63"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Описание ошибки"
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:  #общее кол-во элем-ов во всех стр-цах
          type: integer
          example: 100
        page:   #тек. стр-ца
          type: integer
          example: 1
        limit:   #кол-во элементов на одной странице
          type: integer
          example: 20

    #-------CATEGORY----------------
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор категории
          example: "9b165097-1c9f-4ea3-bef0-e505baa4ff63"
        name:
          type: string
          description: Название категории
    CategoryData:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Название категории
   

    #-----------------------USER-------------------
...
    UserRegister:
      type: object
      required:
        - login
        - email
        - password
        - birthdate
      properties:
        login:
          type: string
          description: Уникальное имя пользователя
        email:
          type: string
          format: email
          description: Email адрес пользователя
        password:
          type: string
          description: Пароль пользователя
        birthdate:
          type: string
          format: date
          description: Дата рождения
    UserLogin:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          description: Уникальное имя пользователя
        password:
          type: string
          description: Пароль пользователя
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер страницы для пагинации
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: LimitParam
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: UUID категории
      schema:
        type: string
        format: uuid
...
    NameSearchQuery:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Поиск по названию
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Произошла внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Конфликт данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация нового пользователя
      description: Регистрирует нового пользователя в системе
      operationId: RegisterUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Аутентификация пользователя
      description: Аутентифицирует пользователя и возвращает JWT-токен
      operationId: LoginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Пользователь с таким логином не найден
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  #---------------------------------USERS---------------------------------------------------
...
  /categories:
    get:
      tags:
        - Categories
      summary: Получить список всех категорий
      description: Возвращает пагинированный список всех категорий
      operationId: GetAllCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        # - $ref: '#/components/parameters/NameSearchQuery'
      responses:
        '200':
          description: Список категорий
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Categories
      summary: Создать новую категорию
      description: Создает новую категорию напитков (только для модераторов и администраторов)
      operationId: AddCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryData'
      responses:
        '201':
          description: Категория успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Категория с таким именем уже существует
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # /categories/{categoryId}:
  #   get:
  #     summary: Получить категорию по ID
  #     description: Возвращает информацию о категории напитков по ее идентификатору
  #     operationId: GetCategoryById
  #     parameters:
  #       - $ref: '#/components/parameters/CategoryId'
  #     responses:
  #       '200':
  #         description: Информация о категории
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Category'
  #       '404':
  #         $ref: '#/components/responses/NotFound'

  #       '500':
  #         $ref: '#/components/responses/InternalServerError'

    # put:
    #   tags:
    #     - Categories
    #   summary: Обновить категорию
    #   description: Обновляет информацию о категории напитков (только для модераторов и администраторов)
    #   operationId: UpdateCategory
    #   security:
    #     - bearerAuth: []
    #   parameters:
    #     - $ref: '#/components/parameters/CategoryId'
    #   requestBody:
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/CategoryData'
    #   responses:
    #     '200':
    #       description: Категория успешно обновлена
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/Category'
    #     '404':
    #       $ref: '#/components/responses/NotFound'
    #     '500':
    #       $ref: '#/components/responses/InternalServerError'

    # delete:
    #   tags:
    #     - Categories
    #   summary: Удалить категорию
    #   description: Удаляет категорию напитков по дентификатору (только для модераторов и администраторов)
    #   operationId: DeleteCategory
    #   security:
    #     - bearerAuth: []
    #   parameters:
    #     - $ref: '#/components/parameters/CategoryId'
    #   responses:
    #     '204':
    #       description: Категория успешно удалена
    #     '404':
    #       $ref: '#/components/responses/NotFound'
    #     '500':
    #       $ref: '#/components/responses/InternalServerError'
  #---------------------------------FAVORITE DRINKS----------------------------------------------------
...